apps/scraper-py/requirements.txt
amqp==5.3.1 ; python_version >= "3.11" and python_version < "4"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4"
anyio==4.9.0 ; python_version >= "3.11" and python_version < "4"
async-timeout==5.0.1 ; python_version >= "3.11" and python_full_version < "3.11.3"
asyncpg==0.30.0 ; python_version >= "3.11" and python_version < "4"
beautifulsoup4==4.13.4 ; python_version >= "3.11" and python_version < "4"
billiard==4.2.1 ; python_version >= "3.11" and python_version < "4"
bota==4.0.152 ; python_version >= "3.11" and python_version < "4"
botasaurus @ git+https://github.com/barbatus/botasaurus.git@9d8356a41d8b94010618f14f8b2b4f7d0264a023 ; python_version >= "3.11" and python_version < "4"
botasaurus-api==4.0.9 ; python_version >= "3.11" and python_version < "4"
botasaurus-driver==4.0.91 ; python_version >= "3.11" and python_version < "4"
botasaurus-humancursor==4.0.83 ; python_version >= "3.11" and python_version < "4"
botasaurus-proxy-authentication==1.0.16 ; python_version >= "3.11" and python_version < "4"
botasaurus-requests==4.0.38 ; python_version >= "3.11" and python_version < "4"
botasaurus-server @ git+https://github.com/barbatus/botasaurus.git@9d8356a41d8b94010618f14f8b2b4f7d0264a023#subdirectory=botasaurus_server ; python_version >= "3.11" and python_version < "4"
bottle==0.13.4 ; python_version >= "3.11" and python_version < "4"
brotli==1.1.0 ; python_version >= "3.11" and python_version < "4"
bs4==0.0.2 ; python_version >= "3.11" and python_version < "4"
cachetools==5.5.2 ; python_version >= "3.11" and python_version < "4"
capsolver-extension-python==1.0.3 ; python_version >= "3.11" and python_version < "4"
casefy==1.1.0 ; python_version >= "3.11" and python_version < "4"
celery==5.5.3 ; python_version >= "3.11" and python_version < "4"
certifi==2025.8.3 ; python_version >= "3.11" and python_version < "4"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "4" and platform_python_implementation == "CPython" and sys_platform == "win32"
charset-normalizer==3.4.2 ; python_version >= "3.11" and python_version < "4"
chrome-extension-python==1.0.4 ; python_version >= "3.11" and python_version < "4"
click-didyoumean==0.3.1 ; python_version >= "3.11" and python_version < "4"
click-plugins==1.1.1.2 ; python_version >= "3.11" and python_version < "4"
click-repl==0.3.0 ; python_version >= "3.11" and python_version < "4"
click==8.2.1 ; python_version >= "3.11" and python_version < "4"
close-chrome==4.0.40 ; python_version >= "3.11" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4" and platform_system == "Windows"
deprecated==1.2.18 ; python_version >= "3.11" and python_version < "4"
durationpy==0.10 ; python_version >= "3.11" and python_version < "4"
et-xmlfile==2.0.0 ; python_version >= "3.11" and python_version < "4"
fastapi==0.116.1 ; python_version >= "3.11" and python_version < "4"
gevent==25.5.1 ; python_version >= "3.11" and python_version < "4"
geventhttpclient==2.3.4 ; python_version >= "3.11" and python_version < "4"
google-auth==2.40.3 ; python_version >= "3.11" and python_version < "4"
greenlet==3.2.3 ; python_version >= "3.11" and python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") or python_version >= "3.11" and platform_python_implementation == "CPython" and python_version < "4"
h11==0.16.0 ; python_version >= "3.11" and python_version < "4"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4"
idna==3.10 ; python_version >= "3.11" and python_version < "4"
javascript-fixes==1.1.29 ; python_version >= "3.11" and python_version < "4"
joblib==1.5.1 ; python_version >= "3.11" and python_version < "4"
kombu==5.5.4 ; python_version >= "3.11" and python_version < "4"
kubernetes==33.1.0 ; python_version >= "3.11" and python_version < "4"
lxml==6.0.0 ; python_version >= "3.11" and python_version < "4"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4"
markdownify==1.1.0 ; python_version >= "3.11" and python_version < "4"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4"
ndjson==0.3.1 ; python_version >= "3.11" and python_version < "4"
numpy==2.3.2 ; python_version >= "3.11" and python_version < "4"
oauthlib==3.3.1 ; python_version >= "3.11" and python_version < "4"
openpyxl==3.1.5 ; python_version >= "3.11" and python_version < "4"
packaging==25.0 ; python_version >= "3.11" and python_version < "4"
prompt-toolkit==3.0.51 ; python_version >= "3.11" and python_version < "4"
psutil==7.0.0 ; python_version >= "3.11" and python_version < "4"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "4"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4" and platform_python_implementation == "CPython" and sys_platform == "win32"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "4"
pydantic==2.11.7 ; python_version >= "3.11" and python_version < "4"
pygments==2.19.2 ; python_version >= "3.11" and python_version < "4"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_version < "4"
pytweening==1.2.0 ; python_version >= "3.11" and python_version < "4"
pyvirtualdisplay==3.0 ; python_version >= "3.11" and python_version < "4"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4"
redis==5.2.1 ; python_version >= "3.11" and python_version < "4"
requests-oauthlib==2.0.0 ; python_version >= "3.11" and python_version < "4"
requests==2.32.4 ; python_version >= "3.11" and python_version < "4"
rich==14.1.0 ; python_version >= "3.11" and python_version < "4"
rsa==4.9.1 ; python_version >= "3.11" and python_version < "4"
setuptools==80.9.0 ; python_version >= "3.11" and python_version < "4"
six==1.17.0 ; python_version >= "3.11" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4"
soupsieve==2.7 ; python_version >= "3.11" and python_version < "4"
sqlalchemy==2.0.42 ; python_version >= "3.11" and python_version < "4"
starlette==0.47.2 ; python_version >= "3.11" and python_version < "4"
typing-extensions==4.14.1 ; python_version >= "3.11" and python_version < "4"
typing-inspection==0.4.1 ; python_version >= "3.11" and python_version < "4"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4"
unidecode==1.4.0 ; python_version >= "3.11" and python_version < "4"
urllib3==2.5.0 ; python_version >= "3.11" and python_version < "4"
uvicorn==0.35.0 ; python_version >= "3.11" and python_version < "4"
vine==5.1.0 ; python_version >= "3.11" and python_version < "4"
wcwidth==0.2.13 ; python_version >= "3.11" and python_version < "4"
websocket-client==1.8.0 ; python_version >= "3.11" and python_version < "4"
wrapt==1.17.2 ; python_version >= "3.11" and python_version < "4"
xlsxwriter==3.2.5 ; python_version >= "3.11" and python_version < "4"
zope-event==5.1.1 ; python_version >= "3.11" and python_version < "4"
zope-interface==7.2 ; python_version >= "3.11" and python_version < "4"
